+++ domain/src/main/java/com/marassu/domain/usecase/GetPetitionAnswerUseCase.kt
@@ -5,6 +5,6 @@ import com.marassu.domain.repository.PetitionAnswerRepository
 class GetPetitionAnswerUseCase(
     private val repository: PetitionAnswerRepository
 ) {
-    suspend fun getPetitionAnswer(petitionId: Int) =
+    suspend fun getPetitionAnswer(petitionId: Long) =
         repository.getPetitionAnswer(petitionId)
 }
